[Week5] Rewrite parser, add support for functions and implement Shunting-yard algorithm for infix -> postfix

+++ src/main/java/me/mikey/challenges/week4/interpreter/expressions/types/commands/DecrCommand.java
@@ -8,6 +8,7 @@ import me.mikey.challenges.week4.interpreter.vm.BBVirtualMachine;
 /**
  * Created by Michael on 27/10/16.
  */
+
 public class DecrCommand extends BBCommand {
     public DecrCommand(Token command, BBArgList argList) {
         super(command, argList);
@@ -16,6 +17,6 @@ public class DecrCommand extends BBCommand {
     @Override
     public void execute(BBVirtualMachine vm) {
         Token variable = argList.getArgList().get(0);
-        vm.setVariable(variable.getData(), vm.getVariableValue(variable.getData(), Integer.class) - 1);
+        vm.setVariable(variable.getData(), vm.getVariableValue(variable.getData(), Double.class) - 1);
     }
 }
